Add topcoder rank list parser

+++ trunk/src/main/java/cn/edu/uestc/acmicpc/web/rank/TrainingRankListUser.java
@@ -6,7 +6,7 @@ public class TrainingRankListUser {
   public TrainingUserDto[] users;
   public Integer rank;
   public Integer solved;
-  public Integer score;
+  public Double score;
   public Integer successfulHack;
   public Integer unsuccessfulHack;
   public Long penalty;
